+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/request/inbound/InboundItemReqVo.java
@@ -4,6 +4,8 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * @author HuangLong
  * @date 2018/12/24
@@ -72,19 +74,19 @@ public class InboundItemReqVo {
     private Long taxRate;
 
     @ApiModelProperty("含税单价")
-    private Long price;
+    private BigDecimal price;
 
     @ApiModelProperty("采购数量")
     private Long num;
 
     @ApiModelProperty("采购含税总价")
-    private Long totalPrice;
+    private BigDecimal totalPrice;
 
     @ApiModelProperty("实际到货数量")
     private Long actualNum;
 
     @ApiModelProperty("实际到货含税总价")
-    private Long actualTotalPrice;
+    private BigDecimal actualTotalPrice;
 
     @ApiModelProperty("规格编码")
     private String specCode;
@@ -99,10 +101,10 @@ public class InboundItemReqVo {
     private String saleUnitName;
 
     @ApiModelProperty("不含税单价")
-    private Long noTaxPrice;
+    private BigDecimal noTaxPrice;
 
     @ApiModelProperty("不含税总价")
-    private Long noTaxTotalPrice;
+    private BigDecimal noTaxTotalPrice;
 
     @ApiModelProperty("主单位和采购单位间的转换系数")
     private Long convertNum;
