+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/request/returngoods/SupplyReturnOrderInfoReqVO.java
@@ -5,6 +5,7 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -112,16 +113,16 @@ public class SupplyReturnOrderInfoReqVO {
     private String consignee;
 
     @ApiModelProperty("运费")
-    private Long deliverAmount;
+    private BigDecimal deliverAmount;
 
     @ApiModelProperty("商品总金额")
-    private Long productTotalAmount;
+    private BigDecimal productTotalAmount;
 
     @ApiModelProperty("优惠额度")
-    private Long discountAmount;
+    private BigDecimal discountAmount;
 
     @ApiModelProperty("订单金额")
-    private Long orderAmount;
+    private BigDecimal orderAmount;
 
     @ApiModelProperty("商品数量")
     private Long productNum;
@@ -149,7 +150,7 @@ public class SupplyReturnOrderInfoReqVO {
     private Date operatorTime;
 
     @ApiModelProperty("活动优惠")
-    private Long activityDiscount;
+    private BigDecimal activityDiscount;
 
     @ApiModelProperty("重量")
     private Long weight;
