+++ src/main/java/com/aiqin/bms/scmp/api/purchase/domain/pojo/order/OrderInfoItem.java
@@ -4,6 +4,8 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @ApiModel("订单商品信息")
 @Data
 public class OrderInfoItem {
@@ -56,13 +58,13 @@ public class OrderInfoItem {
     private String batchNumber;
 
     @ApiModelProperty("分销单价")
-    private Long price;
+    private BigDecimal price;
 
     @ApiModelProperty("数量")
     private Long num;
 
     @ApiModelProperty("分销总价")
-    private Long amount;
+    private BigDecimal amount;
 
     @ApiModelProperty("活动分摊")
     private Long activityApportionment;
@@ -92,22 +94,22 @@ public class OrderInfoItem {
     private String companyCode;
 
     @ApiModelProperty("渠道单价")
-    private Long channelUnitPrice;
+    private BigDecimal channelUnitPrice;
 
     @ApiModelProperty("渠道总价")
-    private Long totalChannelPrice;
+    private BigDecimal totalChannelPrice;
 
     @ApiModelProperty("实际渠道单价")
-    private Long actualChannelUnitPrice;
+    private BigDecimal actualChannelUnitPrice;
 
     @ApiModelProperty("实际渠道总价")
-    private Long actualTotalChannelPrice;
+    private BigDecimal actualTotalChannelPrice;
 
     @ApiModelProperty("实际分销总价")
-    private Long actualAmount;
+    private BigDecimal actualAmount;
 
     @ApiModelProperty("实际分销单价")
-    private Long actualPrice;
+    private BigDecimal actualPrice;
 
     @ApiModelProperty("税率")
     private Long tax;
